MINOR: remove unused MeteredKeyValueStore
