+++ streams/src/main/java/org/apache/kafka/streams/state/internals/KeyValueStoreBuilder.java
@@ -27,7 +27,6 @@ public class KeyValueStoreBuilder<K, V> extends AbstractStoreBuilder<K, V, KeyVa
 
     private final KeyValueBytesStoreSupplier storeSupplier;
 
-
     public KeyValueStoreBuilder(final KeyValueBytesStoreSupplier storeSupplier,
                                 final Serde<K> keySerde,
                                 final Serde<V> valueSerde,
@@ -39,11 +38,12 @@ public class KeyValueStoreBuilder<K, V> extends AbstractStoreBuilder<K, V, KeyVa
 
     @Override
     public KeyValueStore<K, V> build() {
-        return new MeteredKeyValueBytesStore<>(maybeWrapCaching(maybeWrapLogging(storeSupplier.get())),
-                                               storeSupplier.metricsScope(),
-                                               time,
-                                               keySerde,
-                                               valueSerde);
+        return new MeteredKeyValueStore<>(
+            maybeWrapCaching(maybeWrapLogging(storeSupplier.get())),
+            storeSupplier.metricsScope(),
+            time,
+            keySerde,
+            valueSerde);
     }
 
     private KeyValueStore<Bytes, byte[]> maybeWrapCaching(final KeyValueStore<Bytes, byte[]> inner) {
